TestProteinProperties.cpp: making most tests pass.
There was an unfortunate aspect in the translation step in one of the tests, the C# to C++ completely removed
the actual configuration of the Modifications, and because of that the test could not pass. Had to go back to the C# version
to see what the test actually tried to accomplish.

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>