TestChromatogram.cpp: documentation of the issue
I want to use this commit to document what is the problem with TestChromatogram.cpp.
I will not continue working on this for now, since I do not think that it is used by
MetaMorpheus, but might revisit it later.

When creating a Chromatogram object, peakList is allocated, but not set correctly, i.e. the Peak constructor for every pair of (x,y) should be called. See Spectra/Spectrum.h for his.
Trouble is that there are multiple classes derived from Spectrum (MzSpectrum, GeneratedMzSpectrum), and multiple Peaks (MzPeak, GeneratedMzPeak, ChromatographicPeak) and the issue is occuring on each of them.

Furthermore, I am a bit concerned about the inheritance aspects of all of these classes, it might require a re-design of this code section overall. Note also, that it makes a nervous that the template used for the Peak object is a pointer, since the actual class elements of the obect are not available because of that.

Will revisit if this turns out to be relevant.

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>