fix the segfaults in SampleLoadModWithLongMotif
two things:

xmlTextReaderReadString(*xml) might return a nullptr, which makes the constructor to a std::string segfault as well as xmlFree();
So need to protect code by an  if ( delXmlChar1 != nullptr ) {} block

In addition, needed the allocation of some temporary variables. The code converter generates often code that is roughly as follows

SequenceVariation tempVar(....);
privateSequenceVariation.push_back(&tempVar);

this is incorrect, since tempVar will be destroyed once the function exits, since it is a local variable. Need to change all of these occurences to

auto tempVar2 = new SequenceVariation(...)
privateSequenceVariation.push_back(tempVar);

(the same for other elements and datatypes).

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>