update Math.h and Assert.h to MzLibMath.h and MzLibAssert.h
this change is triggered by an attempt to compile MzLib on a Mac OS.
Mac Os file systems are case insensitive (in contrary to Linux file systems),
and Assert.h was pulled in incorrectly at locations where assert.h ( a compiler
header) was actually required.

This change makes the names now unique to avoid the conflict on case-insensitive file systems.

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>