PeptideWithSetModifications.h: revamp the Serialization and Deserialization
this is part of an operation that will be required for MPI later, and therefore we have to be able to store
the serialized data not just in a file, but also in a char buffer for communication.

Add a new Test for this, which will be complemented by other classes later.

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>