IsotopicDistribution: fix the GetDistribution function
a lot of memory cleanup, and some minor fixes.
The most significant change is the order in which elements are processed
in GetDistribution. There seems to be a difference between the order
presented by the C# Hashset and the C++ unordered_map, and the order
seems to matter. And yes, I know, both C# and C++ say explicitely that you
cannot rely on the order when iterating over an unordered_map etc.

Signed-off-by: Edgar Gabriel <egabriel@central.uh.edu>